+++ src/base/BaseJsxPlugin/v4.ts
@@ -1,23 +1,18 @@
-import type { AppType, UserDefinedOptions, IMangleOptions } from '@/types'
+import type { AppType, UserDefinedOptions, InternalUserDefinedOptions, IBaseWebpackPlugin } from '@/types'
 import type { Compiler } from 'webpack4'
-import { styleHandler } from '@/postcss'
-import { createInjectPreflight } from '@/postcss/preflight'
-import { jsxHandler } from '@/jsx'
 import { getOptions } from '@/defaults'
-import { pluginName } from '@/shared'
+import { pluginName, NS } from '@/constants'
 import { ConcatSource, Source } from 'webpack-sources'
-import type { IBaseWebpackPlugin } from '@/interfaces'
-import { NS } from '@/constants'
 import path from 'path'
 import { getGroupedEntries } from '@/base/shared'
-import ClassGenerator from '@/mangle/classGenerator'
+// import ClassGenerator from '@/mangle/classGenerator'
 /**
  * @issue https://github.com/sonofmagic/weapp-tailwindcss-webpack-plugin/issues/5
  */
 export class BaseJsxWebpackPluginV4 implements IBaseWebpackPlugin {
-  options: Required<UserDefinedOptions>
+  options: InternalUserDefinedOptions
   appType: AppType
-  classGenerator?: ClassGenerator
+  // classGenerator?: ClassGenerator
   static NS = NS
   constructor (options: UserDefinedOptions = { framework: 'react' }, appType: AppType) {
     this.options = getOptions(options)
@@ -25,30 +20,14 @@ export class BaseJsxWebpackPluginV4 implements IBaseWebpackPlugin {
   }
 
   apply (compiler: Compiler) {
-    const {
-      jsMatcher,
-      mainCssChunkMatcher,
-      replaceUniversalSelectorWith,
-      framework,
-      cssPreflight,
-      customRuleCallback,
-      cssPreflightRange,
-      disabled,
-      onLoad,
-      onUpdate,
-      onEnd,
-      onStart,
-      mangle,
-      loaderOptions
-    } = this.options
+    const { jsMatcher, mainCssChunkMatcher, framework, disabled, onLoad, onUpdate, onEnd, onStart, loaderOptions, styleHandler, jsxHandler } = this.options
     if (disabled) {
       return
     }
-    if (mangle) {
-      this.classGenerator = new ClassGenerator(mangle as IMangleOptions)
-    }
+    // if (mangle) {
+    //   this.classGenerator = new ClassGenerator(mangle as IMangleOptions)
+    // }
 
-    const cssInjectPreflight = createInjectPreflight(cssPreflight)
     const isReact = framework === 'react'
     const loader = path.resolve(__dirname, `${NS}.js`)
     onLoad()
@@ -68,7 +47,7 @@ export class BaseJsxWebpackPluginV4 implements IBaseWebpackPlugin {
           const rule = {
             loader, // Path to loader
             options: {
-              framework,
+              jsxHandler,
               // classGenerator,
               write: loaderOptions.jsxRename
             }
@@ -93,7 +72,7 @@ export class BaseJsxWebpackPluginV4 implements IBaseWebpackPlugin {
           // }
 
           const rawSource = originalSource.source().toString()
-          const { code } = jsxHandler(rawSource, framework)
+          const { code } = jsxHandler(rawSource)
           const source = new ConcatSource(code)
           compilation.updateAsset(file, source)
           onUpdate(file, rawSource, code)
@@ -101,19 +80,15 @@ export class BaseJsxWebpackPluginV4 implements IBaseWebpackPlugin {
       }
       if (Array.isArray(groupedEntries.css)) {
         for (let i = 0; i < groupedEntries.css.length; i++) {
-          let classGenerator
+          // let classGenerator
           const [file, originalSource] = groupedEntries.css[i]
-          if (this.classGenerator && this.classGenerator.isFileIncluded(file)) {
-            classGenerator = this.classGenerator
-          }
+          // if (this.classGenerator && this.classGenerator.isFileIncluded(file)) {
+          //   classGenerator = this.classGenerator
+          // }
           const rawSource = originalSource.source().toString()
           const css = styleHandler(rawSource, {
-            isMainChunk: mainCssChunkMatcher(file, this.appType),
-            cssInjectPreflight,
-            customRuleCallback,
-            cssPreflightRange,
-            replaceUniversalSelectorWith,
-            classGenerator
+            isMainChunk: mainCssChunkMatcher(file, this.appType)
+            // classGenerator
           })
           const source = new ConcatSource(css)
           compilation.updateAsset(file, source)