+++ src/base/escape.ts
@@ -1,14 +1,12 @@
-import { MappingChars2StringEntries as entries } from '@/dic'
-
-// https://github.com/sindresorhus/escape-string-regexp
-function escapeStringRegexp (str: string) {
-  if (typeof str !== 'string') {
-    throw new TypeError('Expected a string')
-  }
-  return str.replace(/[|\\{}()[\]^$+*?.]/g, '\\$&').replace(/-/g, '\\x2d')
-}
-
-export function escape (selectors: string, raw?: boolean) {
+import { MappingChars2StringEntries as defaultEntries } from '@/dic'
+import { escapeStringRegexp } from '@/reg'
+/**
+ * @description 转义正则
+ * @param selectors
+ * @param raw
+ * @returns
+ */
+export function escape (selectors: string, raw: boolean = false, entries: [string, string][] = defaultEntries) {
   let res = selectors
   for (let i = 0; i < entries.length; i++) {
     const [searchValue, replaceValue] = entries[i]