+++ src/loader/jsx-rename-loader.ts
@@ -1,7 +1,6 @@
 import type * as webpack from 'webpack'
 import { getOptions } from 'loader-utils'
 import type { JsxRenameLoaderOptions } from '@/types'
-import { jsxHandler } from '@/jsx'
 import { mkfileSync } from './util'
 import path from 'path'
 // "jsx", "flow", "typescript"
@@ -10,11 +9,11 @@ export default function loader (this: webpack.LoaderContext<JsxRenameLoaderOptio
   this.cacheable && this.cacheable()
 
   // @ts-ignore
-  const config: JsxRenameLoaderOptions = getOptions(this)
-  if (config.write) {
-    const t = path.resolve(config.write.dir!, '.' + this.resource.replace(this.context, '') + '.tmp')
+  const { jsxHandler, write }: JsxRenameLoaderOptions = getOptions(this)
+  if (write) {
+    const t = path.resolve(write.dir!, '.' + this.resource.replace(this.context, '') + '.tmp')
     mkfileSync(t, content)
   }
-  const { code } = jsxHandler(content, config.framework)
+  const { code } = jsxHandler(content)
   return code
 }