+++ src/postcss/plugin.ts
@@ -1,5 +1,5 @@
 import type { PluginCreator, Plugin } from 'postcss'
-import type { InternalPostcssOptions, StyleHandlerOptions } from '@/types'
+import type { InternalPostcssOptions, IStyleHandlerOptions } from '@/types'
 import { transformSync } from './selectorParser'
 import { getOptions } from '@/defaults'
 import { commonChunkPreflight } from './mp'
@@ -13,7 +13,7 @@ const plugin: PostcssWeappTailwindcssRenamePlugin = (options: InternalPostcssOpt
   const { classGenerator } = options
   const { cssPreflight, cssPreflightRange, customRuleCallback, replaceUniversalSelectorWith } = mergedOptions
   const cssInjectPreflight = createInjectPreflight(cssPreflight)
-  const opts: StyleHandlerOptions = {
+  const opts: IStyleHandlerOptions = {
     cssInjectPreflight,
     cssPreflightRange,
     isMainChunk: true,