+++ src/postcss/selectorParser.ts
@@ -2,9 +2,9 @@ import selectorParser from 'postcss-selector-parser'
 import { internalCssSelectorReplacer } from './shared'
 import type { SyncProcessor } from 'postcss-selector-parser'
 import type { Rule } from 'postcss'
-import type { StyleHandlerOptions } from '@/types'
+import type { IStyleHandlerOptions } from '@/types'
 
-const createTransform = (rule: Rule, options: StyleHandlerOptions) => {
+const createTransform = (rule: Rule, options: IStyleHandlerOptions) => {
   const replaceFlag = options.replaceUniversalSelectorWith !== false
   const classGenerator = options.classGenerator
 
@@ -22,7 +22,7 @@ const createTransform = (rule: Rule, options: StyleHandlerOptions) => {
       }
 
       if (selector.type === 'class') {
-        selector.value = internalCssSelectorReplacer(selector.value)
+        selector.value = internalCssSelectorReplacer(selector.value, options.escapeEntries)
         if (classGenerator && selector.value) {
           let ignore = false
           const prev = rule.prev()
@@ -42,11 +42,11 @@ const createTransform = (rule: Rule, options: StyleHandlerOptions) => {
   return transform
 }
 
-const getTransformer = (rule: Rule, options: StyleHandlerOptions) => {
+const getTransformer = (rule: Rule, options: IStyleHandlerOptions) => {
   return selectorParser(createTransform(rule, options))
 }
 
-export const transformSync = (rule: Rule, options: StyleHandlerOptions) => {
+export const transformSync = (rule: Rule, options: IStyleHandlerOptions) => {
   const transformer = getTransformer(rule, options)
 
   return transformer.transformSync(rule, {